+++ extract_msg/__main__.py
@@ -2,6 +2,7 @@ __all__ = [
     'main',
 ]
 
+
 import os
 import sys
 import traceback